requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking pyproject.toml
#
anyio==3.6.2
    # via starlette
asgiref==3.6.0
    # via opentelemetry-instrumentation-asgi
backoff==2.2.1
    # via opentelemetry-exporter-otlp-proto-http
certifi==2022.12.7
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via uvicorn
deprecated==1.2.13
    # via opentelemetry-api
fastapi==0.88.0
    # via predictor (pyproject.toml)
googleapis-common-protos==1.57.0
    # via opentelemetry-exporter-otlp-proto-http
h11==0.14.0
    # via uvicorn
idna==3.4
    # via
    #   anyio
    #   requests
opentelemetry-api==1.15.0
    # via
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-sdk
opentelemetry-distro==0.36b0
    # via predictor (pyproject.toml)
opentelemetry-exporter-otlp-proto-http==1.15.0
    # via predictor (pyproject.toml)
opentelemetry-instrumentation==0.36b0
    # via
    #   opentelemetry-distro
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi==0.36b0
    # via opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-fastapi==0.36b0
    # via predictor (pyproject.toml)
opentelemetry-proto==1.15.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.15.0
    # via
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.36b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-sdk
opentelemetry-util-http==0.36b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
protobuf==4.21.12
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pydantic==1.10.4
    # via fastapi
requests==2.28.1
    # via opentelemetry-exporter-otlp-proto-http
sniffio==1.3.0
    # via anyio
starlette==0.22.0
    # via fastapi
typing-extensions==4.4.0
    # via
    #   opentelemetry-sdk
    #   pydantic
urllib3==1.26.13
    # via requests
uvicorn==0.20.0
    # via predictor (pyproject.toml)
wrapt==1.14.1
    # via
    #   deprecated
    #   opentelemetry-instrumentation

# The following packages are considered to be unsafe in a requirements file:
# setuptools
